Fixed code structure. Created folder torchFastText for wrapper class / public lib.
