+++ src/torchFastText/__init__.py
@@ -1,0 +1,3 @@
+"""
+Init script.
+"""
